Adding global space and gauge. TOBEDONE: Implement go_to_R and go_to_k

+++ ci-test/test_fft.cpp
@@ -11,9 +11,9 @@
 */
 int main()
 {
-    int Npoints = 1000;
+    size_t Npoints = 1000;
     mdarray<std::complex<double>,2> Input({1,Npoints});
-    std::vector<int> dimensions = {Npoints};
+    std::vector<int> dimensions = {int(Npoints)};
     
     double i0 = 0;
     double a = 1./double(Npoints);
